refactor(Line & Sections): add Sections and move logic about section to Sections class from Line entity
